requirements.txt
streamlit>=1.24.0
tensorflow>=2.16.1
pillow>=9.5.0
numpy>=1.23.5
matplotlib>=3.7.1
scikit-learn>=1.2.2
transformers>=4.30.2